fix(local): preserve GPT-5.6 picker metadata
