+++ scripts/source-file-size-baseline.json
@@ -9,12 +9,12 @@
   "src/cli/app/AppCoordinator.tsx": 5085,
   "src/cli/app/AppView.tsx": 1724,
   "src/cli/app/use-approval-flow.ts": 1163,
-  "src/cli/app/use-configuration-handlers.ts": 1384,
+  "src/cli/app/use-configuration-handlers.ts": 1383,
   "src/cli/app/use-conversation-loop.ts": 2935,
   "src/cli/app/use-submit-handler.ts": 3768,
   "src/cli/components/AgentSelector.tsx": 1270,
   "src/cli/components/InputRich.tsx": 2219,
-  "src/cli/components/ModelSelector.tsx": 1299,
+  "src/cli/components/ModelSelector.tsx": 1204,
   "src/cli/components/ProviderSelector.tsx": 1676,
   "src/cli/components/ToolCallMessageRich.tsx": 1109,
   "src/cli/helpers/accumulator.ts": 1596,
