+++ src/cli/components/ModelSelector.tsx
@@ -11,9 +11,6 @@ import {
 import {
   CHATGPT_FAST_SERVICE_TIER,
   getChatGptFastRegistryHandleForModelHandle,
-  getLocalModelLabel,
-  getModelInfo,
-  isLocalModelHandle,
   models,
   normalizeModelHandleForRegistry,
 } from "@/agent/model";
@@ -25,6 +22,18 @@ import {
 } from "@/providers/byok-providers";
 import { settingsManager } from "@/settings-manager";
 import { colors } from "./colors";
+import {
+  baseHandleForByokAlias,
+  filterModelsByAvailabilityForSelector,
+  includeUnknownBackendHandleInRecommended,
+  labelForBackendModel,
+  type ModelSelectorSelection,
+  registryHandleForBackendModel,
+  registryHandleForByokAlias,
+  toByokSelectorModel,
+  toSelectorModelForHandle,
+  type UiModel,
+} from "./model-selector-helpers";
 import { OverlayShell } from "./OverlayShell";
 import { TabBar } from "./TabBar";
 import { Text } from "./Text";
@@ -40,11 +49,22 @@ type ModelCategory =
   | "server-recommended"
   | "server-all";
 
-const CHATGPT_OAUTH_BASE_PROVIDER = "openai-codex";
-const CHATGPT_LABEL_SUFFIX_PATTERN = /\s+\(ChatGPT\)$/;
-
 // Re-export for consumers that import from ModelSelector
 export { buildByokProviderAliases, isByokHandleForSelector };
+export type {
+  ModelSelectorSelection,
+  UiModel,
+} from "./model-selector-helpers";
+export {
+  filterModelsByAvailabilityForSelector,
+  includeUnknownBackendHandleInRecommended,
+  labelForBackendModel,
+  labelForChatGPTByokAlias,
+  registryHandleForBackendModel,
+  registryHandleForByokAlias,
+  toByokSelectorModel,
+  toSelectorModelForHandle,
+} from "./model-selector-helpers";
 
 export function usesBackendModelCatalog(
   isSelfHosted?: boolean,
@@ -102,170 +122,6 @@ export function getModelCategories(
     return ["recents", ...base];
   }
   return base;
-}
-
-export type UiModel = {
-  id: string;
-  handle: string;
-  label: string;
-  description: string;
-  registryHandle?: string;
-  isDefault?: boolean;
-  isFeatured?: boolean;
-  free?: boolean;
-  updateArgs?: Record<string, unknown>;
-};
-
-export type ModelSelectorSelection = Pick<
-  UiModel,
-  "id" | "handle" | "label" | "description" | "registryHandle" | "updateArgs"
->;
-
-export function labelForChatGPTByokAlias(
-  label: string,
-  handle: string,
-  byokProviderAliases: Record<string, string>,
-): string {
-  const slashIndex = handle.indexOf("/");
-  if (slashIndex === -1) return label;
-
-  const providerAlias = handle.slice(0, slashIndex);
-  if (byokProviderAliases[providerAlias] !== CHATGPT_OAUTH_BASE_PROVIDER) {
-    return label;
-  }
-
-  return label.replace(CHATGPT_LABEL_SUFFIX_PATTERN, ` (${providerAlias})`);
-}
-
-export function baseHandleForByokAlias(
-  handle: string,
-  byokProviderAliases: Record<string, string>,
-): string {
-  const slashIndex = handle.indexOf("/");
-  if (slashIndex === -1) return handle;
-
-  const provider = handle.slice(0, slashIndex);
-  const model = handle.slice(slashIndex + 1);
-  const baseProvider = byokProviderAliases[provider];
-
-  if (baseProvider) {
-    return `${baseProvider}/${model}`;
-  }
-  return handle;
-}
-
-export function registryHandleForByokAlias(
-  handle: string,
-  byokProviderAliases: Record<string, string>,
-): string {
-  const baseHandle = baseHandleForByokAlias(handle, byokProviderAliases);
-  return normalizeModelHandleForRegistry(baseHandle) ?? baseHandle;
-}
-
-export function registryHandleForBackendModel(
-  handle: string,
-  providerType?: string,
-): string {
-  const normalizedHandle = normalizeModelHandleForRegistry(handle) ?? handle;
-  if (models.some((model) => model.handle === normalizedHandle)) {
-    return normalizedHandle;
-  }
-
-  if (providerType === "chatgpt_oauth") {
-    const slashIndex = handle.indexOf("/");
-    if (slashIndex > 0) {
-      const directHandle = `openai/${handle.slice(slashIndex + 1)}`;
-      if (models.some((model) => model.handle === directHandle)) {
-        return directHandle;
-      }
-    }
-  }
-
-  return normalizedHandle;
-}
-
-export function labelForBackendModel(
-  label: string,
-  providerType?: string,
-): string {
-  return providerType === "chatgpt_oauth" ? `${label} (ChatGPT)` : label;
-}
-
-export function toByokSelectorModel(
-  staticModel: UiModel,
-  handle: string,
-  byokProviderAliases: Record<string, string>,
-  updateArgs?: Record<string, unknown>,
-): UiModel {
-  const resolvedUpdateArgs =
-    updateArgs ??
-    (staticModel.updateArgs as Record<string, unknown> | undefined);
-
-  return {
-    ...staticModel,
-    id: handle,
-    handle,
-    registryHandle: registryHandleForByokAlias(handle, byokProviderAliases),
-    label: labelForChatGPTByokAlias(
-      staticModel.label,
-      handle,
-      byokProviderAliases,
-    ),
-    updateArgs: resolvedUpdateArgs,
-  };
-}
-
-export function toSelectorModelForHandle(handle: string): UiModel {
-  const registryHandle = normalizeModelHandleForRegistry(handle) ?? handle;
-  const modelInfo = getModelInfo(registryHandle);
-  if (modelInfo) {
-    return {
-      id: handle,
-      handle,
-      registryHandle,
-      label: modelInfo.label,
-      description: modelInfo.description ?? "",
-      updateArgs: modelInfo.updateArgs as Record<string, unknown> | undefined,
-    };
-  }
-  return {
-    id: handle,
-    handle,
-    label: getLocalModelLabel(handle),
-    description: "",
-  };
-}
-
-const API_GATED_MODEL_HANDLES = new Set([
-  "letta/auto",
-  "letta/auto-fast",
-  "letta/glm",
-]);
-
-export function includeUnknownBackendHandleInRecommended(
-  handle: string,
-): boolean {
-  const registryHandle = normalizeModelHandleForRegistry(handle) ?? handle;
-  return isLocalModelHandle(registryHandle);
-}
-
-export function filterModelsByAvailabilityForSelector<
-  T extends { handle: string },
->(
-  typedModels: T[],
-  availableHandles: Set<string> | null,
-  allApiHandles: string[],
-): T[] {
-  if (availableHandles === null) {
-    return typedModels.filter((m) => {
-      if (!API_GATED_MODEL_HANDLES.has(m.handle)) {
-        return true;
-      }
-      return allApiHandles.includes(m.handle);
-    });
-  }
-
-  return typedModels.filter((m) => availableHandles.has(m.handle));
 }
 
 interface ModelSelectorProps {
@@ -838,6 +694,12 @@ export function ModelSelector({
       // When availableHandles is non-null, skip unavailable models
       if (availableHandles !== null && !availableHandles.has(handle)) continue;
 
+      if (backendModelCatalog) {
+        const backendModel = modelsForBackendHandle(handle, true)[0];
+        if (backendModel) resolved.push(backendModel);
+        continue;
+      }
+
       // Try to resolve to a static model with label/description
       const staticModel = pickPreferredStaticModel(toBaseHandle(handle));
       if (staticModel) {
@@ -866,7 +728,9 @@ export function ModelSelector({
     return resolved;
   }, [
     availableHandles,
+    backendModelCatalog,
     byokProviderAliases,
+    modelsForBackendHandle,
     pickPreferredStaticModel,
     toBaseHandle,
     withProviderTypeMetadata,
