+++ src/models.json
@@ -1367,6 +1367,19 @@
       }
     },
     {
+      "id": "gpt-5.6-sol-plus-pro-max",
+      "handle": "chatgpt-plus-pro/gpt-5.6-sol",
+      "label": "GPT-5.6 Sol (ChatGPT)",
+      "description": "GPT-5.6 Sol (max reasoning) via ChatGPT Plus/Pro",
+      "updateArgs": {
+        "reasoning_effort": "max",
+        "verbosity": "low",
+        "context_window": 272000,
+        "max_output_tokens": 128000,
+        "parallel_tool_calls": true
+      }
+    },
+    {
       "id": "gpt-5.6-terra-plus-pro-none",
       "handle": "chatgpt-plus-pro/gpt-5.6-terra",
       "label": "GPT-5.6 Terra (ChatGPT)",
@@ -1425,6 +1438,19 @@
       "description": "GPT-5.6 Terra (extra-high reasoning) via ChatGPT Plus/Pro",
       "updateArgs": {
         "reasoning_effort": "xhigh",
+        "verbosity": "low",
+        "context_window": 272000,
+        "max_output_tokens": 128000,
+        "parallel_tool_calls": true
+      }
+    },
+    {
+      "id": "gpt-5.6-terra-plus-pro-max",
+      "handle": "chatgpt-plus-pro/gpt-5.6-terra",
+      "label": "GPT-5.6 Terra (ChatGPT)",
+      "description": "GPT-5.6 Terra (max reasoning) via ChatGPT Plus/Pro",
+      "updateArgs": {
+        "reasoning_effort": "max",
         "verbosity": "low",
         "context_window": 272000,
         "max_output_tokens": 128000,
@@ -1501,9 +1527,22 @@
       "id": "gpt-5.6-sol-xhigh",
       "handle": "openai/gpt-5.6-sol",
       "label": "GPT-5.6 Sol",
-      "description": "OpenAI's most capable GPT-5.6 model (max reasoning)",
+      "description": "OpenAI's most capable GPT-5.6 model (extra-high reasoning)",
       "updateArgs": {
         "reasoning_effort": "xhigh",
+        "verbosity": "medium",
+        "context_window": 1050000,
+        "max_output_tokens": 128000,
+        "parallel_tool_calls": true
+      }
+    },
+    {
+      "id": "gpt-5.6-sol-max",
+      "handle": "openai/gpt-5.6-sol",
+      "label": "GPT-5.6 Sol",
+      "description": "OpenAI's most capable GPT-5.6 model (max reasoning)",
+      "updateArgs": {
+        "reasoning_effort": "max",
         "verbosity": "medium",
         "context_window": 1050000,
         "max_output_tokens": 128000,
@@ -1567,9 +1606,22 @@
       "id": "gpt-5.6-terra-xhigh",
       "handle": "openai/gpt-5.6-terra",
       "label": "GPT-5.6 Terra",
-      "description": "GPT-5.6 Terra (max reasoning)",
+      "description": "GPT-5.6 Terra (extra-high reasoning)",
       "updateArgs": {
         "reasoning_effort": "xhigh",
+        "verbosity": "medium",
+        "context_window": 1050000,
+        "max_output_tokens": 128000,
+        "parallel_tool_calls": true
+      }
+    },
+    {
+      "id": "gpt-5.6-terra-max",
+      "handle": "openai/gpt-5.6-terra",
+      "label": "GPT-5.6 Terra",
+      "description": "GPT-5.6 Terra (max reasoning)",
+      "updateArgs": {
+        "reasoning_effort": "max",
         "verbosity": "medium",
         "context_window": 1050000,
         "max_output_tokens": 128000,
@@ -1633,9 +1685,22 @@
       "id": "gpt-5.6-luna-xhigh",
       "handle": "openai/gpt-5.6-luna",
       "label": "GPT-5.6 Luna",
-      "description": "GPT-5.6 Luna (max reasoning)",
+      "description": "GPT-5.6 Luna (extra-high reasoning)",
       "updateArgs": {
         "reasoning_effort": "xhigh",
+        "verbosity": "medium",
+        "context_window": 1050000,
+        "max_output_tokens": 128000,
+        "parallel_tool_calls": true
+      }
+    },
+    {
+      "id": "gpt-5.6-luna-max",
+      "handle": "openai/gpt-5.6-luna",
+      "label": "GPT-5.6 Luna",
+      "description": "GPT-5.6 Luna (max reasoning)",
+      "updateArgs": {
+        "reasoning_effort": "max",
         "verbosity": "medium",
         "context_window": 1050000,
         "max_output_tokens": 128000,
